Time: 0 ms (100.00%) | Memory: 89.4 MB (56.90%) - LeetSync

+++ 2358-number-of-ways-to-split-array/number-of-ways-to-split-array.cpp
@@ -3,11 +3,11 @@ public:
     int waysToSplitArray(vector<int>& nums) {
         int n = nums.size();
         long long sum = 0;
-        long long count = 0;
-        long long totalsum = accumulate(nums.begin(),nums.end(),(long long)0);
-        for(int i = 0;i<n-1;i++){
+        int count = 0;
+        long long totalsum = accumulate(nums.begin(), nums.end(), (long long)0);
+        for (int i = 0; i < n - 1; i++) {
             sum += nums[i];
-            if(sum >= totalsum - sum){
+            if (sum >= totalsum - sum) {
                 count++;
             }
         }
